More refactoring to make re-useable components for IO based entities
