+++ src/components/critters/CritterDetails.tsx
@@ -1,22 +1,12 @@
-import IIOEntity from '../../interfaces/IIOEntity';
 import ResourceChips from '../resources/ResourceChips';
 import DetailsBase from '../ui/DetailsBase';
 import DetailsSection from '../ui/DetailsSection';
-import { FC, memo, useMemo } from 'react';
+import { IRecordDetailsProps } from '../ui/IOGridCard';
+import { FC, memo } from 'react';
 
-interface IProps {
-  critter: IIOEntity;
-}
-
-export const CritterDetails: FC<IProps> = memo(({ critter }) => {
-  const imgUrl = useMemo(
-    () =>
-      `/images/critter/${critter.name.toLowerCase().split(' ').join('-')}.png`,
-    [critter.name],
-  );
-
+export const CritterDetails: FC<IRecordDetailsProps> = memo(({ critter }) => {
   return (
-    <DetailsBase title={critter.name} imgUrl={imgUrl}>
+    <DetailsBase title={critter.name} imgUrl={critter.imgUrl}>
       <DetailsSection title="Inputs">
         <ResourceChips ios={critter.inputs} />
       </DetailsSection>
